Changed the 'bitpreservation database' into 'archive database' for deploy. Also fixed the unit tests.
Split up the url for this external database. The url now consists of 4 elements: baseUrl, machine, port and directory. It should only be access through the function dk.netarkivet.archive.arcrepositoryadmin.DBConnect.getArchiveUrl(), since it combines the different settings into the correct url. If a specific url is wanted, just place it in the baseUrl and set the other settings (machine, port and directory) to empty.
Also fixed the relevant unit tests, and all the functions handling the database.