Created post-processing for batchjobs. All that is needed is that the batchjob overrides the function: public boolean postProcess(InputStream input, OutputStream output)
If it returns false, then the file at the end of the input stream is returned. This file is the standard concatenation of the result files.
If on the other had it returns true, then the file at the end of the output stream is returned as the result of the batchjob.